docs : added page for archive notes
