+++ src/Pages/Account/Account.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import './Account.css';
+
+export default function Account() {
+  return (
+    <>
+    <div className="notes-container">
+        <h1>Account Information</h1>
+    </div>
+    </>
+  );
+}
