+++ src/components/Sidebar/Sidebar.css
@@ -1,0 +1,29 @@
+.sidebar{
+    display: flex;
+    flex-direction: column;
+    line-height: 2.2;
+    min-width: 12rem;
+    text-align: initial;
+}
+
+.sidebar-link{
+    font-size: 1.2rem;
+    font-weight: 500;
+    cursor: pointer;
+    text-decoration: none;
+}
+
+.create-note-btn{
+    background-color: var(--primary-color);
+    color: var(--white-color);
+    padding: 8px 14px;
+    border: none;
+    margin-top: 1rem;
+    cursor: pointer;
+    font-size: 1rem;
+    transition: .4s all ease-in;
+}
+
+.create-note-btn:hover{
+    background-color: var(--primary-dark-color);
+}
